+++ test/plugin/rack_file_test.rb
@@ -1,6 +1,7 @@
 require "test_helper"
 require "shrine/plugins/rack_file"
 require "down"
+require "delegate"
 
 describe Shrine::Plugins::RackFile do
   before do
@@ -16,6 +17,30 @@ describe Shrine::Plugins::RackFile do
 
   it "enables assignment of Rack file hashes" do
     uploaded_file = @uploader.upload(@rack_file)
+
+    assert_equal "image",      uploaded_file.read
+    assert_equal 5,            uploaded_file.size
+    assert_equal "image.jpg",  uploaded_file.original_filename
+    assert_equal "image/jpeg", uploaded_file.mime_type
+  end
+
+  it "accepts hash with indifferent access like objects" do
+    hash_like_class = Class.new(Hash) do
+      def initialize(hash)
+        @hash = hash.inject({}) { |h, (key, value)| h.update(key.to_s => value) }
+      end
+
+      def key?(name)
+        @hash.key?(name.to_s)
+      end
+
+      def [](name)
+        @hash[name.to_s]
+      end
+    end
+
+    hash_like_object = hash_like_class.new(@rack_file)
+    uploaded_file = @uploader.upload(hash_like_object)
 
     assert_equal "image",      uploaded_file.read
     assert_equal 5,            uploaded_file.size